requirements.txt
lxml
PyPDF2
python-dotenv
google-generativeai
pinecone-clien
fuzzywuzzy
python-Levenshtein 
fitz 